+++ rectools/models/nn/transformer_data_preparator.py
@@ -114,6 +114,7 @@ class SessionEncoderDataPreparatorBase:
         train_min_user_interactions: int = 2,
         n_negatives: tp.Optional[int] = None,
         get_val_mask_func: tp.Optional[tp.Callable] = None,
+        **kwargs: tp.Any,
     ) -> None:
         """TODO"""
         self.item_id_map: IdMap
@@ -236,7 +237,7 @@ class SessionEncoderDataPreparatorBase:
 
         Returns
         -------
-        Optional(Dataset)
+        Optional(DataLoader)
             Validation dataloader.
         """
         if self.val_interactions is None:
@@ -252,8 +253,15 @@ class SessionEncoderDataPreparatorBase:
         )
         return val_dataloader
 
-    def get_dataloader_recommend(self, dataset: Dataset) -> DataLoader:
-        """TODO"""
+    def get_dataloader_recommend(self, dataset: Dataset, batch_size: int) -> DataLoader:
+        """
+        Construct recommend dataloader from processed dataset.
+
+        Returns
+        -------
+        DataLoader
+            Recommend dataloader.
+        """
         # Recommend dataloader should return interactions sorted by user ids.
         # User ids here are internal user ids in dataset.interactions.df that was prepared for recommendations.
         # Sorting sessions by user ids will ensure that these ids will also be correct indexes in user embeddings matrix
@@ -261,7 +269,7 @@ class SessionEncoderDataPreparatorBase:
         sequence_dataset = SequenceDataset.from_interactions(interactions=dataset.interactions.df, sort_users=True)
         recommend_dataloader = DataLoader(
             sequence_dataset,
-            batch_size=self.batch_size,
+            batch_size=batch_size,
             collate_fn=self._collate_fn_recommend,
             num_workers=self.dataloader_num_workers,
             shuffle=False,